requirements.txt
aioredis==1.3.1
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.3
async-timeout==3.0.1
billiard==3.6.3.0
celery==4.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
configparser==5.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
crayons==0.4.0
cryptography==2.8
defusedxml==0.7.0rc1
Django==3.0.4
django-braces==1.14.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.8
django-nine==0.2.3
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
entrypoints==0.3
flake8==3.7.9
hiredis==1.1.0
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
Pillow==7.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
redis==3.3.10
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
selenium==3.141.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.2.5
webdriver-manager==2.5.1
